refactor: 정보 아이콘 경로 수정
